authentication/requirements.txt
flask
flask-cors
typing
PyJWT
cryptography